+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Add new `jackson3` artifact for Jackson 3 ( #2878 )
 - Deprecate for removal installPlugins plugin discovery. It's too easy to get yourself into trouble.
 - Support TYPE_USE `@Nullable` annotations like JSpecify ( #2899, thanks @protocol7 ! )